fix: package 역할 재조정
