+++ packages/components/src/theme/index.ts
@@ -1,2 +1,2 @@
-export { DJ_DEFAULT_THEME } from './theme.constants';
+export { DJ_THEME_WITH_COMPONENT } from './theme.constants';
 export { createColorByColorScheme, createColorAlpha } from './theme.utils';
