+++ packages/css/src/index.ts
@@ -1,4 +1,4 @@
+export * from './theme';
 export * from './base';
 export * from './cssVar';
 export * from './system-props';
-export { systemProps, stylePropList } from './system';
