+++ packages/css/src/system-props/index.ts
@@ -1,7 +1,2 @@
-export { default as aliases } from './aliases';
-export { default as background } from './background';
-export { default as color } from './color';
-export { default as size } from './size';
-export { default as typography } from './typography';
-export { shouldTransformToVarFunc } from './scales';
+export { systemProps, stylePropList } from './base';
 export type * from './types';
