+++ packages/css/src/theme/constants.ts
@@ -1,3 +1,6 @@
+import foundation from './foundation';
+import semanticTokens from './semanticTokens';
+
 export const THEME = {
   KEY: 'dj',
   DEFAULT_KEY: '__default',
@@ -17,3 +20,9 @@ export type PseudoKeys = keyof typeof TOKEN_PSEUDO_CLASSES;
 export type PseudoAliases = (typeof TOKEN_ALIASES)[keyof typeof TOKEN_ALIASES];
 
 export const pseudoKeys = Object.values(TOKEN_ALIASES);
+
+export const DJ_DEFAULT_THEME = {
+  key: THEME.KEY,
+  ...foundation,
+  semanticTokens,
+} as const;
