+++ packages/css/src/theme/index.ts
@@ -1,4 +1,3 @@
-export * from './base';
 export * from './constants';
 export { default as foundation } from './foundation';
 export { default as semanticTokens } from './semanticTokens';
