+++ packages/styled/package.json
@@ -11,7 +11,6 @@
     "dist/**"
   ],
   "scripts": {
-    "test": "jest",
     "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
     "lint": "TIMING=1 eslint --ext .ts,.tsx src/",
     "build": "tsup src/index.ts --clean --format esm,cjs --dts --minify"
