+++ packages/styled/src/index.ts
@@ -1,6 +1,6 @@
 export { keyframes } from '@emotion/react';
+export { DJ_DEFAULT_THEME } from '@danji/css';
 export * from './factory';
-export * from './providers';
-export * from './theme';
+export { Global, useTheme, ThemeProvider, forwardRef } from './providers';
 export { default as CssReset } from './reset';
 export type * from './base.types';
