+++ packages/styled/src/providers.types.ts
@@ -1,21 +1,3 @@
 export interface CssVariablesProps {
   selector?: string;
 }
-
-export type NormalToken = string | number;
-
-export type NormalizedToken = {
-  _light: NormalToken;
-};
-
-export type SemanticToken = NormalizedToken & {
-  _dark?: NormalToken;
-};
-
-export type DesignToken = NormalToken | SemanticToken | DesignTokenObject;
-
-export type DesignTokens = Record<string, DesignToken>;
-
-export interface DesignTokenObject {
-  [key: string]: DesignToken;
-}
