+++ packages/utilities/src/index.ts
@@ -1,5 +1,6 @@
 export * from './arrayUtils';
 export * from './assertion';
+export * from './colorUtils';
 export * from './localStorage';
 export * from './objectUtils';
 export * from './sharedUtils';
